refactor: extracts helper methods to another file
