+++ __tests__/integration/walletservice_facade.test.ts
@@ -4,7 +4,7 @@ import config from '../../src/config';
 import { loggers } from './utils/logger.util';
 import HathorWalletServiceWallet from '../../src/wallet/wallet';
 import Network from '../../src/models/network';
-import { CreateTokenTransaction, MemoryStore, Output, Storage } from '../../src';
+import { CreateTokenTransaction, Output, Storage } from '../../src';
 import {
   FULLNODE_NETWORK_NAME,
   FULLNODE_URL,
@@ -19,36 +19,25 @@ import {
 import { decryptData } from '../../src/utils/crypto';
 import walletUtils from '../../src/utils/wallet';
 import { delay } from './utils/core.util';
-import { TxNotFoundError, UtxoError, WalletRequestError } from '../../src/errors';
+import { UtxoError, WalletRequestError } from '../../src/errors';
 import { NATIVE_TOKEN_UID } from '../../lib/constants';
 import { GetAddressesObject } from '../../lib/wallet/types';
+import {
+  buildWalletInstance,
+  emptyWallet,
+  generateNewWalletAddress,
+  poolForTx,
+} from './helpers/service-facade.helper';
+import { GenesisWalletServiceHelper } from './helpers/genesis-wallet.helper';
 
 // Set base URL for the wallet service API inside the privatenet test container
 config.setWalletServiceBaseUrl('http://localhost:3000/dev/');
 config.setWalletServiceBaseWsUrl('ws://localhost:3001/');
 
 /** Genesis Wallet, used to fund all tests */
-const gWallet: HathorWalletServiceWallet = buildWalletInstance({
-  words: WALLET_CONSTANTS.genesis.words,
-}).wallet;
+const gWallet: HathorWalletServiceWallet = GenesisWalletServiceHelper.getSingleton();
 /** Wallet instance used in tests */
 let wallet: HathorWalletServiceWallet;
-const emptyWallet = {
-  words:
-    'buddy kingdom scorpion device uncover donate sense false few leaf oval illegal assume talent express glide above brain end believe abstract will marine crunch',
-  addresses: [
-    'WkHNZyrKNusTtu3EHfvozEqcBdK7RoEMR7',
-    'WivGyxDjWxijcns3hpGvEJKhjR9HMgFzZ5',
-    'WXQSeMcNt67hVpmgwYqmYLsddgXeGYP4mq',
-    'WTMH3NQs8YXyNguqwLyqoTKDFTfkJLxMzX',
-    'WTUiHeiajtt1MXd1Jb3TEeWUysfNJfig35',
-    'WgzZ4MNcuX3sBgLC5Fa6dTTQaoy4ccLdv5',
-    'WU6UQCnknGLh1WP392Gq6S69JmheS5kzZ2',
-    'WX7cKt38FfgKFWFxSa2YzCWeCPgMbRR98h',
-    'WZ1ABXsuwHHfLzeAWMX7RYs5919LPBaYpp',
-    'WUJjQGb4SGSLh44m2JdgAR4kui8mTPb8bK',
-  ],
-};
 const walletWithTxs = {
   words:
     'bridge balance milk impact love orchard achieve matrix mule axis size hip cargo rescue truth stable setup problem nerve fit million manage harbor connect',
@@ -135,133 +124,14 @@ const pinCode = '123456';
 /** Default password to simplify the tests */
 const password = 'testpass';
 
-/**
- * Builds a HathorWalletServiceWallet instance with a wallet seed words
- * @param enableWs - Whether to enable websocket connection (default: false)
- * @param words - The 24 words to use for the wallet (default: empty wallet)
- * @param passwordForRequests - The password that will be returned by the mocked requestPassword function (default: 'test-password')
- * @returns The wallet instance along with its store and storage for eventual mocking/spying
- */
-function buildWalletInstance({
-  enableWs = false,
-  words = emptyWallet.words,
-  passwordForRequests = 'test-password',
-} = {}) {
-  const walletData = { words };
-  const network = new Network(NETWORK_NAME);
-  const requestPassword = jest.fn().mockResolvedValue(passwordForRequests);
-
-  const store = new MemoryStore();
-  const storage = new Storage(store);
-  const newWallet = new HathorWalletServiceWallet({
-    requestPassword,
-    seed: walletData.words,
-    network,
-    storage,
-    enableWs, // Disable websocket for integration tests
-  });
-
-  return { wallet: newWallet, store, storage };
-}
-
-/**
- * Polls the wallet for a transaction by its ID until found or max attempts reached
- * @param walletForPolling - The wallet instance to poll
- * @param txId - The transaction ID to look for
- * @returns The transaction object if found
- * @throws Error if the transaction is not found after max attempts
- */
-async function poolForTx(walletForPolling: HathorWalletServiceWallet, txId: string) {
-  const maxAttempts = 10;
-  const delayMs = 1000; // 1 second
-  let attempts = 0;
-
-  while (attempts < maxAttempts) {
-    try {
-      const tx = await walletForPolling.getTxById(txId);
-      if (tx) {
-        loggers.test!.log(`Pooling for ${txId} took ${attempts + 1} attempts`);
-        return tx;
-      }
-    } catch (error) {
-      // If the error is of type TxNotFoundError, we continue polling
-      if (!(error instanceof TxNotFoundError)) {
-        throw error; // Re-throw unexpected errors
-      }
-    }
-    attempts++;
-    await delay(delayMs);
-  }
-  throw new Error(`Transaction ${txId} not found after ${maxAttempts} attempts`);
-}
-
-async function generateNewWalletAddress() {
-  const newWords = walletUtils.generateWalletWords();
-  const { wallet: newWallet } = buildWalletInstance({ words: newWords });
-  await newWallet.start({ pinCode, password });
-
-  const addresses: string[] = [];
-  for (let i = 0; i < 10; i++) {
-    addresses.push((await newWallet.getAddressAtIndex(i))!);
-  }
-
-  return {
-    words: newWords,
-    addresses,
-  };
-}
-
-async function sendFundTx(
-  address: string,
-  amount: bigint,
-  destinationWallet?: HathorWalletServiceWallet
-) {
-  const fundTx = await gWallet.sendTransaction(address, amount, {
-    pinCode,
-  });
-
-  // Ensure the transaction was sent from the Genesis perspective
-  await poolForTx(gWallet, fundTx.hash!);
-
-  // Ensure the destination wallet is also aware of the transaction
-  if (destinationWallet) {
-    await poolForTx(destinationWallet, fundTx.hash!);
-  }
-
-  return fundTx;
-}
-
 beforeAll(async () => {
   console.log(`${JSON.stringify(await generateNewWalletAddress(), null, 2)}`);
-
-  let isServerlessReady = false;
-  const startTime = Date.now();
-
-  // Pool for the serverless app to be ready.
-  const delayBetweenRequests = 3000;
-  const lambdaTimeout = 30000;
-  while (isServerlessReady) {
-    try {
-      // Executing a method that does not depend on the wallet being started,
-      // but that ensures the Wallet Service Lambdas are receiving requests
-      await gWallet.getVersionData();
-      isServerlessReady = true;
-    } catch (e) {
-      // Ignore errors, serverless app is probably not ready yet
-      loggers.test!.log('Ws-Serverless not ready yet, retrying in 3 seconds...');
-    }
-
-    // Timeout after 2 minutes
-    if (Date.now() - startTime > lambdaTimeout) {
-      throw new Error('Ws-Serverless did not become ready in time');
-    }
-    await delay(delayBetweenRequests);
-  }
-  await gWallet.start({ pinCode, password });
+  await GenesisWalletServiceHelper.poolForServerlessAvailable();
+  await GenesisWalletServiceHelper.start();
 });
 
 afterAll(async () => {
-  await gWallet.stop({ cleanStorage: true });
+  await GenesisWalletServiceHelper.stop();
 });
 
 describe('start', () => {
@@ -703,13 +573,11 @@ describe('basic transaction methods', () => {
     });
 
     it('should create a new token without any custom options', async () => {
-      const fundTx = await sendFundTx(customTokenWallet.addresses[0], 10n);
-
       ({ wallet } = buildWalletInstance({
         words: customTokenWallet.words,
       }));
       await wallet.start({ pinCode, password });
-      await poolForTx(wallet, fundTx.hash!);
+      await GenesisWalletServiceHelper.injectFunds(customTokenWallet.addresses[0], 10n, wallet);
 
       const createTokenTx = (await wallet.createNewToken(tokenName, tokenSymbol, tokenAmount, {
         pinCode,
@@ -1066,13 +934,11 @@ describe('basic transaction methods', () => {
     });
 
     it('should create token with all outputs to another wallet', async () => {
-      const fundTx = await sendFundTx(customTokenWallet.addresses[0], 10n);
-
       ({ wallet } = buildWalletInstance({
         words: customTokenWallet.words,
       }));
       await wallet.start({ pinCode, password });
-      await poolForTx(wallet, fundTx.hash!);
+      await GenesisWalletServiceHelper.injectFunds(customTokenWallet.addresses[0], 10n, wallet);
 
       // Assign external addresses from multipleTokensWallet (starting from index 9 going backwards)
       const destinationAddress = multipleTokensWallet.addresses[9]; // Token destination
@@ -1294,7 +1160,10 @@ describe('websocket events', () => {
       // Add your assertions here after triggering the event
     });
 
-    const sendTransaction = await sendFundTx(walletWithTxs.addresses[0], 10n);
+    const sendTransaction = await GenesisWalletServiceHelper.injectFunds(
+      walletWithTxs.addresses[0],
+      10n
+    );
     expect(sendTransaction.hash).toBeDefined();
 
     // Wait up to 3 times, 2 seconds each, for events to arrive
@@ -1549,7 +1418,7 @@ describe('getUtxos, getUtxosForAmount, getAuthorityUtxos', () => {
     await utxosTestWallet.start({ pinCode, password });
 
     // Fund the wallet with multiple transactions to create various UTXOs
-    await sendFundTx(utxosWallet.addresses[0], 100n, utxosTestWallet);
+    await GenesisWalletServiceHelper.injectFunds(utxosWallet.addresses[0], 100n, utxosTestWallet);
 
     // Create additional UTXOs by sending to different addresses
     const fundTx2 = await utxosTestWallet.sendTransaction(utxosWallet.addresses[1], 20n, {
